Registrator/requirements.txt
certifi==2018.4.16
chardet==3.0.4
docker-py==1.10.6
docker-pycreds==0.3.0
idna==2.6
requests==2.18.4
six==1.11.0
urllib3==1.22
websocket-client==0.48.0